+++ tango/integrations/beaker/executor.py
@@ -1,12 +1,11 @@
+import json
 import logging
 import os
-import tempfile
 import threading
 import time
 import uuid
 import warnings
 from abc import abstractmethod
-from pathlib import Path
 from typing import Dict, List, NamedTuple, Optional, Sequence, Set, Tuple, Union
 
 from beaker import (
@@ -17,6 +16,8 @@ from beaker import (
     DatasetNotFound,
     Digest,
     EnvVar,
+    Experiment,
+    ExperimentNotFound,
     ExperimentSpec,
     JobFailedError,
     JobTimeoutError,
@@ -42,7 +43,7 @@ from tango.step_info import GitMetadata
 from tango.version import VERSION
 from tango.workspace import Workspace
 
-from .common import Constants
+from .common import Constants, get_client
 
 logger = logging.getLogger(__name__)
 
@@ -64,9 +65,9 @@ class ResourceAssignment(NamedTuple):
     Resources assigned to a step.
     """
 
-    cluster: str
+    cluster: Union[str, List[str]]
     """
-    The cluster to use to execute the step.
+    The cluster(s) to use to execute the step.
     """
 
     resources: TaskResources
@@ -126,7 +127,6 @@ class SimpleBeakerScheduler(BeakerScheduler):
         super().__init__()
         self.clusters = clusters
         self.priority = priority
-        self._latest_clusters_used: List[str] = []
         if not self.clusters:
             raise ConfigurationError("At least one cluster is required in 'clusters'")
 
@@ -138,44 +138,9 @@ class SimpleBeakerScheduler(BeakerScheduler):
             memory=step_resources.memory,
             shared_memory=step_resources.shared_memory,
         )
-        cluster_to_use = self._ensure_cluster(task_resources)
-        if cluster_to_use is None:
-            raise ResourceAssignmentError()
-
-        # Move cluster to the end of `self._latest_clusters_used`
-        try:
-            self._latest_clusters_used.remove(cluster_to_use)  # type: ignore
-        except ValueError:
-            pass
-        self._latest_clusters_used.append(cluster_to_use)  # type: ignore
-
         return ResourceAssignment(
-            cluster=cluster_to_use, resources=task_resources, priority=self.priority
+            cluster=self.clusters, resources=task_resources, priority=self.priority
         )
-
-    def _ensure_cluster(self, task_resources: TaskResources) -> Optional[str]:
-        cluster_to_use: Optional[str] = None
-        if not self.clusters:
-            raise ConfigurationError("At least one cluster is required in 'clusters'")
-        elif len(self.clusters) == 1:
-            cluster_to_use = self.clusters[0]
-        else:
-
-            def recency_ranking(cluster_name: str):
-                try:
-                    return self._latest_clusters_used.index(cluster_name)
-                except ValueError:
-                    return -1
-
-            available_clusters = sorted(
-                self.beaker.cluster.filter_available(task_resources, *self.clusters),
-                key=lambda x: (x.queued_jobs, recency_ranking(x.cluster.full_name)),
-            )
-
-            if available_clusters:
-                cluster_to_use = available_clusters[0].cluster.full_name
-
-        return cluster_to_use
 
 
 @Executor.register("beaker")
@@ -329,20 +294,6 @@ class BeakerExecutor(Executor):
 
     """
 
-    GITHUB_TOKEN_SECRET_NAME: str = "TANGO_GITHUB_TOKEN"
-
-    BEAKER_TOKEN_SECRET_NAME: str = "BEAKER_TOKEN"
-
-    RESULTS_DIR: str = "/tango/output"
-
-    ENTRYPOINT_DIR: str = "/tango/entrypoint"
-
-    ENTRYPOINT_FILENAME: str = "entrypoint.sh"
-
-    INPUT_DIR: str = "/tango/input"
-
-    STEP_GRAPH_FILENAME: str = "config.json"
-
     DEFAULT_BEAKER_IMAGE: str = "ai2/conda"
     """
     The default image. Used if neither ``beaker_image`` nor ``docker_image`` are set.
@@ -410,7 +361,7 @@ class BeakerExecutor(Executor):
 
         super().__init__(workspace, include_package=include_package, parallelism=parallelism)
 
-        self.beaker = Beaker.from_env(default_workspace=beaker_workspace, session=True, **kwargs)
+        self.beaker = get_client(beaker_workspace=beaker_workspace, **kwargs)
         self.beaker_image = beaker_image
         self.docker_image = docker_image
         self.datasets = datasets
@@ -528,6 +479,7 @@ class BeakerExecutor(Executor):
                                 # A dependency failed or can't be run, so this step can't be run.
                                 not_run[step_name] = ExecutionMetadata()
                                 steps_to_run.discard(step_name)
+                                steps_left_to_run.discard(step)
                             break
                     else:
                         # Dependencies are OK, so we can run this step now.
@@ -570,6 +522,51 @@ class BeakerExecutor(Executor):
 
             return future_done_callback
 
+        last_progress_update = time.monotonic()
+
+        def log_progress():
+            nonlocal last_progress_update
+
+            now = time.monotonic()
+            if now - last_progress_update >= 60 * 2:
+                last_progress_update = now
+
+                waiting_for = [
+                    step_name
+                    for step_name in submitted_steps
+                    if step_name not in failed and step_name not in successful
+                ]
+                if len(waiting_for) > 5:
+                    logger.info(
+                        "Waiting for %d steps...",
+                        len(waiting_for),
+                    )
+                elif len(waiting_for) > 1:
+                    logger.info(
+                        "Waiting for %d steps (%s)...",
+                        len(waiting_for),
+                        "'" + "', '".join(waiting_for) + "'",
+                    )
+                elif len(waiting_for) == 1:
+                    logger.info("Waiting for 1 step ('%s')...", list(waiting_for)[0])
+
+                still_to_run = [
+                    step.name for step in steps_left_to_run if step.name not in submitted_steps
+                ]
+                if len(still_to_run) > 5:
+                    logger.info(
+                        "Still waiting to submit %d more steps...",
+                        len(still_to_run),
+                    )
+                elif len(still_to_run) > 1:
+                    logger.info(
+                        "Still waiting to submit %d more steps (%s)...",
+                        len(still_to_run),
+                        "'" + "', '".join(still_to_run) + "'",
+                    )
+                elif len(still_to_run) == 1:
+                    logger.info("Still waiting to submit 1 more step ('%s')...", still_to_run[0])
+
         update_steps_to_run()
 
         try:
@@ -602,6 +599,8 @@ class BeakerExecutor(Executor):
 
                     # Update the step queue.
                     update_steps_to_run()
+
+                    log_progress()
         except (KeyboardInterrupt, CancellationError):
             if step_futures:
                 cli_logger.warning("Received interrupt, canceling steps...")
@@ -630,37 +629,6 @@ class BeakerExecutor(Executor):
                 "to become available..."
             )
 
-    def execute_sub_graph_for_step(
-        self, step_graph: StepGraph, step_name: str, run_name: Optional[str] = None
-    ) -> ExecutorOutput:
-        self.check_repo_state()
-        while True:
-            try:
-                step = step_graph[step_name]
-                experiment_url = self._execute_sub_graph_for_step(step_graph, step_name)
-                return ExecutorOutput(
-                    successful={
-                        step_name: ExecutionMetadata(
-                            result_location=None
-                            if not step.cache_results
-                            else self.workspace.step_info(step).result_location,
-                            logs_location=experiment_url,
-                        )
-                    }
-                )
-            except ResourceAssignmentError:
-                self._emit_resource_assignment_warning()
-                time.sleep(3.0)
-            except StepFailedError as exc:
-                return ExecutorOutput(
-                    failed={step_name: ExecutionMetadata(logs_location=exc.experiment_url)}
-                )
-            except ExecutorError:
-                return ExecutorOutput(failed={step_name: ExecutionMetadata()})
-            except Exception as exc:
-                log_exception(exc, logger)
-                return ExecutorOutput(failed={step_name: ExecutionMetadata()})
-
     def _check_if_cancelled(self):
         if self._is_cancelled.is_set():
             raise RunCancelled
@@ -685,32 +653,65 @@ class BeakerExecutor(Executor):
             )
             return None
 
-        # Initialize experiment and task spec.
-        experiment_name, spec = self._build_experiment_spec(step_graph, step_name)
-        self._check_if_cancelled()
+        experiment: Optional[Experiment] = None
+        experiment_url: Optional[str] = None
 
-        step.log_starting()
+        # Try to find any existing experiments for this step that are still running.
+        if step.cache_results:
+            for exp in self.beaker.workspace.experiments(
+                match=f"{Constants.STEP_EXPERIMENT_PREFIX}{step.unique_id}-"
+            ):
+                self._check_if_cancelled()
+                try:
+                    latest_job = self.beaker.experiment.latest_job(exp)
+                except (ValueError, ExperimentNotFound):
+                    continue
+                if latest_job is not None and not latest_job.is_done:
+                    experiment = exp
+                    experiment_url = self.beaker.experiment.url(exp)
+                    cli_logger.info(
+                        "[blue]\N{black rightwards arrow} Found existing Beaker experiment [b]%s[/] for "
+                        'step [b]"%s"[/] that is still running...[/]',
+                        experiment_url,
+                        step_name,
+                    )
+                    break
 
-        # Create experiment.
-        experiment = self.beaker.experiment.create(experiment_name, spec)
-        experiment_url = self.beaker.experiment.url(experiment)
-        cli_logger.info(
-            '[blue]\N{black rightwards arrow} Submitted Beaker experiment [b]%s[/] for step [b]"%s"[/]...[/]',
-            experiment_url,
-            step_name,
-        )
+        # Otherwise we submit a new experiment...
+        if experiment is None:
+            # Initialize experiment and task spec.
+            experiment_name, spec = self._build_experiment_spec(step_graph, step_name)
+            self._check_if_cancelled()
 
-        # Follow the experiment and stream the logs until it completes.
+            step.log_starting()
+
+            # Create experiment.
+            experiment = self.beaker.experiment.create(experiment_name, spec)
+            experiment_url = self.beaker.experiment.url(experiment)
+            cli_logger.info(
+                '[blue]\N{black rightwards arrow} Submitted Beaker experiment [b]%s[/] for step [b]"%s"[/]...[/]',
+                experiment_url,
+                step_name,
+            )
+
+        assert experiment is not None
+        assert experiment_url is not None
+
+        # Follow the experiment until it completes.
         try:
             while True:
                 try:
                     self._check_if_cancelled()
                     self.beaker.experiment.wait_for(
-                        experiment, strict=True, quiet=True, timeout=2.0
+                        experiment,
+                        strict=True,
+                        quiet=True,
+                        timeout=31,
+                        poll_interval=30,
                     )
-                    time.sleep(2.0)
                     break
                 except JobTimeoutError:
+                    time.sleep(30)
                     continue
         except (JobFailedError, TaskStoppedError):
             cli_logger.error(
@@ -759,7 +760,7 @@ class BeakerExecutor(Executor):
 
         # Get hash of the local entrypoint source file.
         sha256_hash = hashlib.sha256()
-        contents = read_binary(tango.integrations.beaker, "entrypoint.sh")
+        contents = read_binary(tango.integrations.beaker, Constants.ENTRYPOINT_FILENAME)
         sha256_hash.update(contents)
 
         entrypoint_dataset_name = (
@@ -777,17 +778,16 @@ class BeakerExecutor(Executor):
             # Create it.
             logger.debug(f"Creating entrypoint dataset '{entrypoint_dataset_name}'")
             try:
-                with tempfile.TemporaryDirectory() as tmpdirname:
-                    tmpdir = Path(tmpdirname)
-                    entrypoint_path = tmpdir / "entrypoint.sh"
-                    with open(entrypoint_path, "wb") as entrypoint_file:
-                        entrypoint_file.write(contents)
-                    tmp_entrypoint_dataset = self.beaker.dataset.create(
-                        tmp_entrypoint_dataset_name, entrypoint_path, quiet=True
-                    )
-                    entrypoint_dataset = self.beaker.dataset.rename(
-                        tmp_entrypoint_dataset, entrypoint_dataset_name
-                    )
+                tmp_entrypoint_dataset = self.beaker.dataset.create(
+                    tmp_entrypoint_dataset_name, quiet=True, commit=False
+                )
+                self.beaker.dataset.upload(
+                    tmp_entrypoint_dataset, contents, Constants.ENTRYPOINT_FILENAME, quiet=True
+                )
+                self.beaker.dataset.commit(tmp_entrypoint_dataset)
+                entrypoint_dataset = self.beaker.dataset.rename(
+                    tmp_entrypoint_dataset, entrypoint_dataset_name
+                )
             except DatasetConflict:  # could be in a race with another `tango` process.
                 time.sleep(1.0)
                 entrypoint_dataset = self.beaker.dataset.get(entrypoint_dataset_name)
@@ -810,11 +810,14 @@ class BeakerExecutor(Executor):
     def _ensure_step_graph_dataset(self, step_graph: StepGraph) -> Dataset:
         step_graph_dataset_name = f"{Constants.STEP_GRAPH_DATASET_PREFIX}{str(uuid.uuid4())}"
         try:
-            with tempfile.TemporaryDirectory() as tmpdirname:
-                tmpdir = Path(tmpdirname)
-                path = tmpdir / self.STEP_GRAPH_FILENAME
-                step_graph.to_file(path, include_unique_id=True)
-                dataset = self.beaker.dataset.create(step_graph_dataset_name, path, quiet=True)
+            dataset = self.beaker.dataset.create(step_graph_dataset_name, quiet=True, commit=False)
+            self.beaker.dataset.upload(
+                dataset,
+                json.dumps({"steps": step_graph.to_config(include_unique_id=True)}).encode(),
+                Constants.STEP_GRAPH_FILENAME,
+                quiet=True,
+            )
+            self.beaker.dataset.commit(dataset)
         except DatasetConflict:  # could be in a race with another `tango` process.
             time.sleep(1.0)
             dataset = self.beaker.dataset.get(step_graph_dataset_name)
@@ -861,7 +864,7 @@ class BeakerExecutor(Executor):
             )
 
         # Get cluster, resources, and priority to use.
-        cluster, task_resources, priority = self.scheduler.schedule(step)
+        clusters, task_resources, priority = self.scheduler.schedule(step)
         self._check_if_cancelled()
 
         # Ensure dataset with the entrypoint script exists and get it.
@@ -873,11 +876,11 @@ class BeakerExecutor(Executor):
         self._check_if_cancelled()
 
         # Write the GitHub token secret.
-        self.beaker.secret.write(self.GITHUB_TOKEN_SECRET_NAME, self.github_token)
+        self.beaker.secret.write(Constants.GITHUB_TOKEN_SECRET_NAME, self.github_token)
         self._check_if_cancelled()
 
         # Write the Beaker token secret.
-        self.beaker.secret.write(self.BEAKER_TOKEN_SECRET_NAME, self.beaker.config.user_token)
+        self.beaker.secret.write(Constants.BEAKER_TOKEN_SECRET_NAME, self.beaker.config.user_token)
         self._check_if_cancelled()
 
         # Build Tango command to run.
@@ -887,7 +890,7 @@ class BeakerExecutor(Executor):
             "debug",
             "--called-by-executor",
             "beaker-executor-run",
-            self.INPUT_DIR + "/" + self.STEP_GRAPH_FILENAME,
+            Constants.INPUT_DIR + "/" + Constants.STEP_GRAPH_FILENAME,
             step.name,
             self.workspace.url,
         ]
@@ -906,26 +909,26 @@ class BeakerExecutor(Executor):
         task_spec = (
             TaskSpec.new(
                 step.unique_id,
-                cluster,
                 beaker_image=self.beaker_image,
                 docker_image=self.docker_image,
-                result_path=self.RESULTS_DIR,
-                command=["bash", self.ENTRYPOINT_DIR + "/" + self.ENTRYPOINT_FILENAME],
+                result_path=Constants.RESULTS_DIR,
+                command=["bash", Constants.ENTRYPOINT_DIR + "/" + Constants.ENTRYPOINT_FILENAME],
                 arguments=command,
                 resources=task_resources,
                 datasets=self.datasets,
                 env_vars=self.env_vars,
                 priority=priority,
             )
+            .with_constraint(cluster=[clusters] if isinstance(clusters, str) else clusters)
             .with_env_var(name="TANGO_VERSION", value=VERSION)
-            .with_env_var(name="GITHUB_TOKEN", secret=self.GITHUB_TOKEN_SECRET_NAME)
-            .with_env_var(name="BEAKER_TOKEN", secret=self.BEAKER_TOKEN_SECRET_NAME)
+            .with_env_var(name="GITHUB_TOKEN", secret=Constants.GITHUB_TOKEN_SECRET_NAME)
+            .with_env_var(name="BEAKER_TOKEN", secret=Constants.BEAKER_TOKEN_SECRET_NAME)
             .with_env_var(name="GITHUB_REPO", value=f"{github_account}/{github_repo}")
             .with_env_var(name="GIT_REF", value=git_ref)
             .with_env_var(name="PYTHON_VERSION", value=python_version)
             .with_env_var(name="BEAKER_EXPERIMENT_NAME", value=experiment_name)
-            .with_dataset(self.ENTRYPOINT_DIR, beaker=entrypoint_dataset.id)
-            .with_dataset(self.INPUT_DIR, beaker=step_graph_dataset.id)
+            .with_dataset(Constants.ENTRYPOINT_DIR, beaker=entrypoint_dataset.id)
+            .with_dataset(Constants.INPUT_DIR, beaker=step_graph_dataset.id)
         )
 
         if self.venv_name is not None: