+++ tango/integrations/transformers/__init__.py
@@ -62,6 +62,10 @@ library under the corresponding class from the `torch <torch.html>`_ integration
             transformers::AutoModelForCTC::from_pretrained
             transformers::AutoModelForCausalLM::from_config
             transformers::AutoModelForCausalLM::from_pretrained
+            transformers::AutoModelForDepthEstimation::from_config
+            transformers::AutoModelForDepthEstimation::from_pretrained
+            transformers::AutoModelForDocumentQuestionAnswering::from_config
+            transformers::AutoModelForDocumentQuestionAnswering::from_pretrained
             transformers::AutoModelForImageClassification::from_config
             transformers::AutoModelForImageClassification::from_pretrained
             transformers::AutoModelForImageSegmentation::from_config
@@ -94,10 +98,14 @@ library under the corresponding class from the `torch <torch.html>`_ integration
             transformers::AutoModelForTableQuestionAnswering::from_pretrained
             transformers::AutoModelForTokenClassification::from_config
             transformers::AutoModelForTokenClassification::from_pretrained
+            transformers::AutoModelForVideoClassification::from_config
+            transformers::AutoModelForVideoClassification::from_pretrained
             transformers::AutoModelForVision2Seq::from_config
             transformers::AutoModelForVision2Seq::from_pretrained
             transformers::AutoModelForVisualQuestionAnswering::from_config
             transformers::AutoModelForVisualQuestionAnswering::from_pretrained
+            transformers::AutoModelForZeroShotObjectDetection::from_config
+            transformers::AutoModelForZeroShotObjectDetection::from_pretrained
             transformers::AutoModelWithLMHead::from_config
             transformers::AutoModelWithLMHead::from_pretrained