+++ tests/step_graph_test.py
@@ -103,6 +103,27 @@ class TestStepGraph(TangoTestCase):
                 "b_number": {"type": "ref", "ref": "list", "key": 1},
             },
         }
-        step_graph = StepGraph.from_params(deepcopy(config))
+        step_graph = StepGraph.from_params(deepcopy(config))  # type: ignore[arg-type]
         assert [s.name for s in step_graph["added"].dependencies] == ["list"]
         assert step_graph.to_config() == config
+
+    def test_with_forced_dependencies(self):
+        config = {
+            "some_string": {
+                "type": "string",
+                "result": "I should run second",
+                "step_extra_dependencies": [{"type": "ref", "ref": "other_string"}],
+            },
+            "other_string": {"type": "string", "result": "I should run first"},
+            "added": {
+                "type": "concat_strings",
+                "string1": "Some string:",
+                "string2": {"type": "ref", "ref": "some_string"},
+            },
+        }
+        step_graph = StepGraph.from_params(deepcopy(config))  # type: ignore[arg-type]
+        assert step_graph["some_string"].dependencies == {step_graph["other_string"]}
+        assert step_graph["added"].recursive_dependencies == {
+            step_graph["other_string"],
+            step_graph["some_string"],
+        }