+++ CHANGELOG.md
@@ -7,8 +7,54 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v1.0.2](https://github.com/allenai/tango/releases/tag/v1.0.2) - 2022-11-14
+
+### Changed
+
+- `BeakerScheduler` can now return a list of clusters.
+
+## [v1.0.1](https://github.com/allenai/tango/releases/tag/v1.0.1) - 2022-10-20
+
+### Fixed
+
+- `LightningTrainStep` now can take a `Lazy` model object which results in a gauranteed deterministic hash.
+- Fixed issue where remote `Workspace` implementations like `WandbWorkspace` and `BeakerWorkspace` would use the same local cache regardless of the W&B / Beaker workspace
+  being used.
+- Fixed bug with `TorchEvalStep` when constructing callbacks.
+- Fixed some import error issues caused when an integration is not installed.
+- Fix incorrect reporting of final results in `MulticoreExecutor`.
+
+### Changed
+
+- Wandb step cache retries api call in case of timeout
+- `beaker-py >= 1.11` required.
+
+## [v1.0.0](https://github.com/allenai/tango/releases/tag/v1.0.0) - 2022-10-05
+
 ### Added
 
+- Added `step_extra_dependencies` input field to `Step` class that can be used to force a dependency on another step even if the current step doesn't directly depend on the output of the other step. See [#418](https://github.com/allenai/tango/issues/418) for more context.
+
+### Changed
+
+- `beaker-py >= 1.10` required.
+
+### Fixed
+
+- Long log lines will be soft-wrapped to ensure that links are clickable.
+- Fixed a bug where some workspaces could be left in a bad state if a step's `Format` failed to serialize the step's result in `Workspace.step_finished()`.
+- Sometimes functions and methods end up as arguments to steps, which means we have to hash them. Instead of taking
+  a hash of the function, we now take a hash of the function's module and name.
+- Fixed a bug with the Beaker executor where it would hang at the end of a run if a step failed that is a dependency of another step.
+- Fixed tests to work with new version of transformers.
+- Fixed `Executor.execute_sub_graph_for_step()` to be able to run the step's dependencies in parallel.
+
+
+## [v0.14.0](https://github.com/allenai/tango/releases/tag/v0.14.0) - 2022-09-20
+
+### Added
+
+- Adds a function to modify a Hugging Face transformer with IA3 adaptors
 - Added a `BeakerScheduler` registrable class, specified as the argument `scheduler` to `BeakerExecutor`, which controls the resources assigned to steps ran on Beaker.
   Users can implement their own `BeakerScheduler` subclasses to customize the resource assignment behavior.
 
@@ -26,7 +72,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug where creating a `StepInfo` object from params might result in unnecessary imports.
 - Fixed a bug where canceling the Beaker executor might not work properly.
 - Fixed a bug where the trainer trains too much when `train_epochs` is set and you're using gradient accumulation.
+- Fixed a bug where included modules might not be found when using multiprocessing when they're not on `sys.path` / `PYTHONPATH`.
 - Fixed how the results of uncacheable steps are displayed by `tango run`.
+- Beaker executor won't run duplicate cacheable steps at the same time.
 
 ## [v0.13.0](https://github.com/allenai/tango/releases/tag/v0.13.0) - 2022-09-07